+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-19T07:57:33.807Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-19T10:21:55.585Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")