+++ packages/orq-rc/RELEASES.md
@@ -8009,3 +8009,13 @@ Based on:
 - [python v4.10.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.14 - packages/orq-rc
+
+## 2026-05-19 10:24:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.16 - packages/orq-rc